refactor: move chart types and processors to chart module
- Relocate chart types (Key, NoteKind, PlayerSide, Argb, BgaLayer) to chart/types.rs
- Move processor modules from chart/process/{bms,bmson}.rs to {bms,bmson}/process.rs
- Wrap BMS-specific ChartEvent variants in BmsEvent enum
- Move BaseBpmGenerator impls to respective processor modules
