+++ src/bmson.rs
@@ -31,6 +31,7 @@ pub mod bms_to_bmson;
 pub mod bmson_to_bms;
 pub mod parse;
 pub mod prelude;
+pub mod process;
 pub mod pulse;
 
 use std::{
