+++ tests/chart/bmson/continue_time.rs
@@ -3,6 +3,7 @@
 use gametime::{TimeSpan, TimeStamp};
 
 use bms_rs::bmson::parse_bmson;
+use bms_rs::bmson::prelude::BmsonProcessor;
 use bms_rs::chart::prelude::*;
 
 use super::{MICROSECOND_EPSILON, assert_time_close};
